Move efs-csi-role and cluster-shared-role step implementation to aws.go
